requirements.txt
PyYAML==6.0
SQLAlchemy==1.4.29

## optional
# astroid==2.9.0
# autopep8==1.6.0
# greenlet==1.1.2
# isort==5.10.1
# lazy-object-proxy==1.7.1
# mccabe==0.6.1
# platformdirs==2.4.1
# pycodestyle==2.8.0
# pylint==2.12.2
# toml==0.10.2
# wrapt==1.13.3
